requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
passlib
python-jose
python-dotenv
databases
alembic
httpx
pydantic
pydantic[email]
python-multipart
psycopg2
passlib
bcrypt
websocket
pytest
aiosqlite
pytest-asyncio
pytest-cov